The most significant changes in the code include the renaming and refactoring of classes, the addition of new classes, and the modification of methods and variables. The `Hashmap_PG` class was renamed to `LandmassHashmap_PG` and the `LandMass` class was renamed to `Landmass`. The `LandMassTextureHandler` class was also renamed to `LandmassTextureHandler`. New classes `CollisionManager_PG` and `ShipHashmap_PG` were added to handle collision detection and ship management respectively.
Several methods and variables were updated to accommodate these changes. For instance, the `interactWithLandmasses` method in `LandMassHandler_PG.h` now takes a pointer to `PlayerShip` instead of a reference. The `landMasses` variable in the same file was changed from a vector of `LandMass` pointers to a vector of unique pointers to `Landmass`.

The `EnemyShip_PG.h` file was updated to include two new header files and several new methods and variables. The `GameStateManager_PG.h` file was updated to initialize the `currGameState` variable to `GameState::Start`. The `GlobalTextureHandler_PG.h` file was updated to change the type of `landMassTextures` from `LandMassTextureHandler` to `LandmassTextureHandler`. The `Hashmap_PG.h` file was completely removed.

The `LandMassHandler_PG.h`, `LandMassTextures_PG.h`, `LandMass_PG.h`, `Minimap_PG.h`, `ShipMovementHandler_PG.h`, `World_PG.h`, `shipCannonHandler_PG.h`, `LandMassHandler_PG.cpp`, `LandMass_PG.cpp`, `Minimap_PG.cpp`, and `World_PG.cpp` files were all updated to reflect the changes in class names and method parameters.

The `Hashmap_PG` files were renamed to `LandmassHashmap_PG` in the `SFML Project 1.vcxproj` file. The `CollisionManager_PG`, `LandmassHashmap_PG`, and `ShipHashmap_PG` classes were added to the `SFML Project 1.vcxproj.filters` file and implemented in their respective `.cpp` files.

+++ PirateGame/Resource/Minimap_PG.cpp
@@ -40,7 +40,7 @@ void Minimap::update(float shipRotation) {
             sf::Vector2f(landmasses[i]->getSprite().getGlobalBounds().width / 2, landmasses[i]->getSprite().getGlobalBounds().height / 2);
 
         // Calculate distance from the ship to the landmass
-        float distance = sqrt(pow(shipPos.x - landmassPos.x, 2) + pow(shipPos.y - landmassPos.y, 2));
+        float distance = static_cast<float>(sqrt(pow(shipPos.x - landmassPos.x, 2) + pow(shipPos.y - landmassPos.y, 2)));
 
         if (distance <= visibilityRadius) {
             sf::RectangleShape landmassRect;
